web/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
sentence-transformers==2.2.2
pgvector==0.2.4
